+++ app/scripts/controllers/network/util.js
@@ -6,10 +6,12 @@ const {
   ROPSTEN_CODE,
   RINKEYBY_CODE,
   KOVAN_CODE,
+  WANCHAIN_CODE,
   ROPSTEN_DISPLAY_NAME,
   RINKEBY_DISPLAY_NAME,
   KOVAN_DISPLAY_NAME,
   MAINNET_DISPLAY_NAME,
+  WANCHAIN_DISPLAY_NAME,
 } = require('./enums')
 
 const networkToNameMap = {
@@ -20,6 +22,7 @@ const networkToNameMap = {
   [ROPSTEN_CODE]: ROPSTEN_DISPLAY_NAME,
   [RINKEYBY_CODE]: RINKEBY_DISPLAY_NAME,
   [KOVAN_CODE]: KOVAN_DISPLAY_NAME,
+  [WANCHAIN_CODE]: WANCHAIN_DISPLAY_NAME,
 }
 
 const getNetworkDisplayName = key => networkToNameMap[key]