+++ app/scripts/lib/buy-eth-url.js
@@ -15,7 +15,7 @@ function getBuyEthUrl ({ network, amount, address }) {
   let url
   switch (network) {
     case '1':
-      url = `https://buy.coinbase.com/?code=9ec56d01-7e81-5017-930c-513daa27bb6a&amount=${amount}&address=${address}&crypto_currency=ETH`
+      url = 'https://wanfaucet.net'
       break
 
     case '3':
@@ -29,6 +29,7 @@ function getBuyEthUrl ({ network, amount, address }) {
     case '42':
       url = 'https://github.com/kovan-testnet/faucet'
       break
+
   }
   return url
 }