+++ old-ui/app/app.js
@@ -298,6 +298,8 @@ App.prototype.getNetworkName = function () {
     name = 'Kovan Test Network'
   } else if (providerName === 'rinkeby') {
     name = 'Rinkeby Test Network'
+  } else if (providerName === 'wanchain') {
+    name = 'Wanchain Network'
   } else {
     name = 'Unknown Private Network'
   }