+++ old-ui/app/components/transaction-list-item.js
@@ -5,7 +5,9 @@ const connect = require('react-redux').connect
 
 const EthBalance = require('./eth-balance')
 const addressSummary = require('../util').addressSummary
-const explorerLink = require('etherscan-link').createExplorerLink
+const explorerLink = function (hash, network) {
+  return `http://wanscan.io/tx/${hash}`
+}
 const CopyButton = require('./copyButton')
 const vreme = new (require('vreme'))()
 const Tooltip = require('./tooltip')