+++ old-ui/lib/icon-factory.js
@@ -1,6 +1,6 @@
 var iconFactory
 const isValidAddress = require('ethereumjs-util').isValidAddress
-const toChecksumAddress = require('ethereumjs-util').toChecksumAddress
+const toChecksumAddress = require('wanchain-util').toChecksumAddress
 const contractMap = require('eth-contract-metadata')
 
 module.exports = function (jazzicon) {