+++ test/unit/app/buy-eth-url.spec.js
@@ -2,10 +2,8 @@ const assert = require('assert')
 const getBuyEthUrl = require('../../../app/scripts/lib/buy-eth-url')
 
 describe('', function () {
-  const mainnet = {
+  const wanchain = {
     network: '1',
-    amount: 5,
-    address: '0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc',
   }
   const ropsten = {
     network: '3',
@@ -17,16 +15,9 @@ describe('', function () {
     network: '42',
   }
 
-  it('returns coinbase url with amount and address for network 1', function () {
-    const coinbaseUrl = getBuyEthUrl(mainnet)
-    const coinbase = coinbaseUrl.match(/(https:\/\/buy.coinbase.com)/)
-    const amount = coinbaseUrl.match(/(amount)\D\d/)
-    const address = coinbaseUrl.match(/(address)(.*)(?=&)/)
-
-    assert.equal(coinbase[0], 'https://buy.coinbase.com')
-    assert.equal(amount[0], 'amount=5')
-    assert.equal(address[0], 'address=0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc')
-
+  it('returns wanchainfaucet url for network 1', function () {
+    const wanchainUrl = getBuyEthUrl(wanchain)
+    assert.equal(wanchainUrl, 'https://wanfaucet.net')
   })
 
   it('returns metamask ropsten faucet for network 3', function () {