+++ ui/app/components/tx-view.js
@@ -75,16 +75,19 @@ TxView.prototype.renderHeroBalance = function () {
 }
 
 TxView.prototype.renderButtons = function () {
-  const {selectedToken, showModal, history } = this.props
+//  const {selectedToken, showModal, history } = this.props
+  const {selectedToken, history } = this.props
 
   return !selectedToken
     ? (
       h('div.flex-row.flex-center.hero-balance-buttons', [
+/* No deposit option yet
         h('button.btn-primary.hero-balance-button', {
           onClick: () => showModal({
             name: 'DEPOSIT_ETHER',
           }),
         }, this.context.t('deposit')),
+*/
 
         h('button.btn-primary.hero-balance-button', {
           style: {